updated some files

+++ src/keyword.h
@@ -8,6 +8,8 @@
 #ifndef _LAMB_KEYWORD_H
 #define _LAMB_KEYWORD_H
 
+#include "db.h"
+
 #define LAMB_MAX_KEYWORD 1024
 
 typedef struct {
@@ -15,7 +17,7 @@ typedef struct {
     char *list[LAMB_MAX_KEYWORD];
 } lamb_keyword_t;
 
-int lamb_keyword_get_all(lamb_db_t *db, lamb_keyword_t *keys, size_t size);
+int lamb_keyword_get_all(lamb_db_t *db, lamb_keyword_t *keys, int size);
 bool lamb_keyword_check(lamb_keyword_t *keys, char *content);
 
 #endif
